Revision to Code 2040
If I may suggest, for in the future. The methods that were being asked of me as far as what I was being asked to do weren't that difficult. What was the most trouble for me was figuring out how to send and receive the data with the http request. I'm not sure that the process of sending an receiving different types of data should be this difficult for an application but that may just be my opinion.